Rename Request Test file + Add tests for body and method props
